libyang: update from 2.1.148 to 3.7.8
Update libyang from 2.1.148 to 3.7.8

Only the copyright years changed the LICENSE file in commit:
   b13198971 doc UPDATE libyang v3 transition
so update that checksum.

Note that the prefix 'ly_' was added to header files in:
   8f702ee38 libyang UPDATE add prefix for common and config files
so the multilib header install:append was updated in the recipe.
Packages that depend on libyang, such as frr, still compile.

The ptests all continue to pass.

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>
Signed-off-by: Khem Raj <raj.khem@gmail.com>